refactor: 重构Gpio模块为完全模块化的架构，将接口定义与具体实现分离，极大地提高了代码的可维护性和可移植性。
- ✅ 接口与实现完全分离
- ✅ 支持多种硬件平台
- ✅ 代码组织清晰
- ✅ 易于单元测试
- ✅ 向后兼容

+++ src/Adapter/ContinuityCollector.cpp
@@ -1,4 +1,5 @@
 #include "ContinuityCollector.h"
+#include "../HAL/GpioFactory.h"
 #include <algorithm>
 #include <iomanip>
 #include <map>
@@ -282,7 +283,7 @@ ContinuityCollector::calculateStatistics() const {
     std::sort(sortedPins.begin(), sortedPins.end(),
               [](const auto &a, const auto &b) { return a.second > b.second; });
 
-    for (int i = 0; i < 5 && i < sortedPins.size(); i++) {
+    for (size_t i = 0; i < 5 && i < sortedPins.size(); i++) {
         stats.mostActivePins[i] = sortedPins[i].first;
     }
